% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{char_to_numeric}
\alias{char_to_numeric}
\title{convert character to numeric if possible.}
\usage{
char_to_numeric(x)
}
\arguments{
\item{x}{the list of lists to convert}
}
\description{
If vector cannot be converted to numeric without introducing NA's,
returns original data (unlisted where applicable). Numeric data is
divide by 1000 to convert to decimal.
}
