% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{get_account_info}
\alias{get_account_info}
\title{Get accounts and balances}
\usage{
get_account_info(budget, last_knowledge = NULL, token = NULL)
}
\arguments{
\item{budget}{Either a numeric index or name of budget.}

\item{last_knowledge}{An (optional) integer specifying last result from server such
that only new data is returned.}

\item{token}{API token to use. (May be omitted if YNAB_TOKEN env variable
is set.)}
}
\value{
a dataframe containig account details.
}
\description{
Get accounts and balances
}
