% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets.R
\name{get_budget_id}
\alias{get_budget_id}
\title{Get budget name by index or name}
\usage{
get_budget_id(n, token = NULL)
}
\arguments{
\item{n}{Either a numeric index or name of budget}

\item{token}{API token to use. (May be omitted if YNAB_TOKEN env variable
is set.)}
}
\value{
id (uuid format) of matching budget or NULL if not found
}
\description{
Get budget name by index or name
}
