% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets.R
\name{get_budget_names}
\alias{get_budget_names}
\title{Get vector of budget names}
\usage{
get_budget_names(token = NULL)
}
\arguments{
\item{token}{API token to use. (May be omitted if YNAB_TOKEN env variable
is set.)}
}
\value{
vector of names
}
\description{
Get vector of budget names
}
