% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\name{get_categories}
\alias{get_categories}
\title{Get categories for specified budget}
\usage{
get_categories(budget, last_knowledge = NULL, token = NULL)
}
\arguments{
\item{budget}{the uuid, name, or index of the desired budget}

\item{last_knowledge}{optional "last knowledge" value to fetch only new info}

\item{token}{API token to use. (May be omitted if YNAB_TOKEN env variable
is set.)}
}
\value{
data.frame of categories and related data
}
\description{
Get categories for specified budget
}
