% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\name{get_category_balances}
\alias{get_category_balances}
\title{Get category balances for specified budget}
\usage{
get_category_balances(...)
}
\arguments{
\item{...}{additional arguments to get_categories (budget,
last_knowledge, and token, the last two of which are optional.)}
}
\value{
vector of activities
}
\description{
This is a convenience wrapper for get_categories
}
