% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month.R
\name{get_current_month}
\alias{get_current_month}
\title{Get current month's data}
\usage{
get_current_month(budget, last_knowledge = NULL, token = NULL)
}
\arguments{
\item{budget}{the uuid, name, or index of the desired budget}

\item{last_knowledge}{optional "last knowledge" value to fetch only new info}

\item{token}{API token to use. (May be omitted if YNAB_TOKEN env variable
is set.)}
}
\value{
data.frame of this month's activity
}
\description{
Get current month's data
}
