% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month.R
\name{get_current_month_activity}
\alias{get_current_month_activity}
\title{Get current month's total activity}
\usage{
get_current_month_activity(...)
}
\arguments{
\item{...}{additional parameters passed to get_current_month including
budget, last_knowledge, and token (the last two being optional.)}
}
\value{
vector of categories
}
\description{
This is a convenience wrapper for get_current_month
}
