% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{get_current_transactions}
\alias{get_current_transactions}
\title{Get transactions for current month}
\usage{
get_current_transactions(budget, last_knowledge = NULL, token = NULL)
}
\arguments{
\item{budget}{the uuid, name, or index of the desired budget}

\item{last_knowledge}{optional "last knowledge" value to fetch only new info}

\item{token}{API token to use. (May be omitted if YNAB_TOKEN env variable
is set.)}
}
\value{
vector of categories
}
\description{
Get transactions for current month
}
