% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{l_of_l_to_df}
\alias{l_of_l_to_df}
\title{converts list of lists to a data.frame}
\usage{
l_of_l_to_df(x)
}
\arguments{
\item{x}{the list of lists to convert}
}
\description{
NULLs are converted to NA, numeric values are converted to appropriate
type, and lists are converted to vectors.
}
