% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{put_transaction}
\alias{put_transaction}
\title{Put a transaction (to modify)}
\usage{
put_transaction(budget, transaction, token = NULL)
}
\arguments{
\item{budget}{the uuid, name, or index of the desired budget}

\item{transaction}{the transaction data}

\item{token}{API token to use. (May be omitted if YNAB_TOKEN env variable
is set.)}
}
\value{
http response
}
\description{
Put a transaction (to modify)
}
