% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnab-package.R
\docType{package}
\name{rnab-package}
\alias{rnab}
\alias{rnab-package}
\title{RNAB Package}
\description{
This package provides an interface to the API exposed by You Need A Budget (youneedabudget.com).
}
\section{Functions}{



\link[rnab]{get_budget_id}

\link[rnab]{get_categories}

\link[rnab]{get_account_info}

\link[rnab]{get_budget_id}

\link[rnab]{get_budget_names}

\link[rnab]{get_categories}

\link[rnab]{get_category_actvities}

\link[rnab]{get_category_balances}

\link[rnab]{get_category_goals}

\link[rnab]{get_get_category_ids}

\link[rnab]{get_get_category_names}

\link[rnab]{get_current_month}

\link[rnab]{get_current_month_activities}

\link[rnab]{get_current_month_activity}

\link[rnab]{get_current_transactions}

\link[rnab]{put_transaction}
}

\keyword{internal}
