% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{update_transactions}
\alias{update_transactions}
\title{Import transactions from linked accounts.}
\usage{
update_transactions(budget, token = NULL)
}
\arguments{
\item{budget}{the uuid, name, or index of the desired budget}

\item{token}{API token to use. (May be omitted if YNAB_TOKEN env variable
is set.)}
}
\value{
number of imported transactions.
}
\description{
Import transactions from linked accounts.
}
