% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{yget}
\alias{yget}
\title{Core GET request wrapper}
\usage{
yget(endpoint, last_knowledge = NULL, token = NULL)
}
\arguments{
\item{endpoint}{the API endpoint to request.}

\item{last_knowledge}{the last_knowledge value, if known, to prevent
reloading data previously requested.}

\item{token}{API token to use. (May be omitted if YNAB_TOKEN env variable
is set.)}
}
\value{
data returned by request.
}
\description{
Core GET request wrapper
}
