% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ypost}
\alias{ypost}
\title{Core POST request wrapper}
\usage{
ypost(endpoint, body, token = NULL)
}
\arguments{
\item{endpoint}{the API endpoint to request.}

\item{body}{data to send to the endpoint}

\item{token}{API token to use. (May be omitted if YNAB_TOKEN env variable
is set.)}
}
\value{
response to request (in JSON format)
}
\description{
Core POST request wrapper
}
